Updating code and removing useless code
changed the port of the light sensor because 2 was acting weird. Added some sleeps in the code to make sure the noodle doesn't go flying when it's released. Also removed the whole pom finding code because it was bad and I want to remake it. Also adjusted some of the turns to make them better. Also made the robot turn around right after it grabs the noodle so I don't have to turn when placing the noodle down. :)